+++ tests/testthat/test_dirty.R
@@ -31,3 +31,14 @@ test_that("time column of daily data is treated as Date (#114)", {
   expect_is(z$time, "Date")
 
 })
+
+
+test_that("time column of daily data is survives two way conversion (#137)", {
+  x <- structure(list(time = structure(c(16030, 16031, 16034, 16035,
+  16036), class = "Date"), value = c(18680.35, 18766.53, 18741.95,
+  18759.68, 18812.33)), class = "data.frame", row.names = c(NA,
+  -5L))
+
+  z <- ts_na_omit(ts_tbl(ts_ts(x)))
+  expect_equal(z$time, x$time)
+})